Namespace ContentGenerator under Rails::Generators
Follows Rails conventions for app content generators. No behavior changes.
