+++ lib/generators/rails/content/templates/controller.rb.tt
