+++ lib/generators/rails/content/templates/index.html.erb.tt
