+++ lib/generators/rails/content/templates/model.rb.tt
