+++ lib/generators/rails/content/templates/root.erb.tt
