+++ lib/generators/rails/content/templates/show.html.erb.tt
